Switch to embedded HTTP for IPC
Before I was using a JSON file to maintain state, as well as
to allow communication with PHP. This caused some concurrency
issues when PHP was accessing the file when we wanted to write
to it. The file would not write and PHP would not know of the
update until the states changed again.

Now we launch an embedded HTTP server that can server the
live state of the events as the timer sees them at any time.
This allows us to manage concurrency entirely within our
program throught he use of Interlocks and SpinLocks.

We still write changes out to the same JSON file to save state
in case of restart, but we should be the only ones writting
to that file now.