Fix some exception-handling issues
There was a bug where if the ArenaNet APIs were offline, then calls
to them would result in (null) being returned, which would cause
exceptions to be thrown.

This left the worker in a state where the m_TimerSync variable was
never released, and subsequent calls would never actually update.

This patch wraps all of the worker logic in a try-catch, fast fails
if we don't need to run things, and utilizes logging to let us know
if something is going wrong with our results.