Load old data into new timer structure
Exceptions were being thrown in the worker thread due to new events
being processed but not present in the old timer data structure loaded
by the service at startup.

This new code does a standard reset of the timer data and then inserts
the old timer data only if the new timer data is aware of it.