Initial persistant event logic
Some events (e.g. The Frozen Maw) don't have the next event in the
chain start immediately after the last one finished. If the timer
queries during this time, it thinks that the event is no longer running.

This is preliminary code (i.e.: not tested). If a meta is currently
not running and the previous stage was valid and not the last stage,
it will check if that stage's events are all successful. If they are
it will continue to report as being in that stage under the assumption
that the next stage will be starting shortly.