Added support for champion event detection
Now that champions drop valuable loot, it is useful to know which events
have champions in them. The ChampionEventRequest queries ArenaNet for the
list of event names, as well as the official wiki for the list of known
events that contain champions. Any events that are known to contain
champions using this approach will have their polygon display in blue on
the map.

It may be valuable to add blue-tinted event icons for these events to
further distinguish them in the event that they do not have polygons or
to distinguish them when they are in preparation.