Prevent deadlock if the app is closed rapidly
There was an issue where if the app was closed before the eventing
data was loaded, the child thread that loads the data would be killed
before it could signal that the data was loaded. Since the app waits
for the event processing thread to finish before exiting, this would
produce a problem as the event thread was waiting for the loading thread
to complete.

This change allows the event processing thread to continue if either
data is loaded or we were asked to bail out.