Bring popups to forefront on click
This commit implements functionality that upon click will check if the
pushpin is currently displaying a popup. If it is, then the pushpin will
be placed at the top of the global map so it is brought to the front.

This also has the positive side effect of meaning that items with popups
open will remain visible despite any other settings that may have hidden
them.