Include HtmlAgilityPack v1.4.6 in our solution
The APIServer was having issues with malformed HTML when implementing the
NodesAPI. This change drops use of Linq to XML in favor of the HTML
Agility Pack from codeplex for HTML parsing and XPath.

This change also fixes some bugs with the NodesAPI that were a result of
malformed HTML/HTML changes.