Fix a bug with build ids and timer resets
The build id being persisted to the database was the BuildResponse
object's ToString() result, and not the actual numeric build id. This
caused the timers to always reset when restarting the API server, since
the in-memory build id persisted by the publisher would be wiped, and thus
the faulty id from the database was used.