Remove VS Mono 3.2.3 Profile build requirement
Previously I had manually set up my box to support telling Visual Studio
that I wanted to build using Mono 3.2.3 framework. This setup, however, is
relatively hacky and doesn't promote easily working with the code.

The reason this was done in the first place was to facilitate running the
APIServer under Linux. So now I have instead implemented a 'Mono'
configuration, which when specified only builds the APIServer libraries
and ensures that all framework dlls are pointed to. To build the solution
under mono simply use xbuild like so:

  xbuild /p:Configuration=Mono GuildWars2.sln

This will build the APIServer and associated libraries and place them in
the corresponding bin/Mono directories.