Remove unnecessary source in favor of NuGet
When using the VS Mono Profile hack, it was necessary to avoid the use of
NuGet packages since VS would complain about which framework the DLLs were
compiled against. Now that we are letting Mono do it, this patch removes
the source-compiled versions of those packages in favor or using NuGet to
manage them.