Add basic support for the new world boss schedules
This change adds a new MetaEventSchedule class which defines the current
world boss schedule, as well as an API to expose the next time each meta
in the schedule will spawn in order.

Currently it's a bit hacky and hasn't been tested yet. It makes use of the
same MetaEventStatus structure for exposing times, which will likely need
to be changed and done more properly.